chore: replace panel component with CollectionListFilter and fix type FilterChoiceComponent
